+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Yii Framework 2 apidoc extension Change Log
 - Enh #349: Add inheritance of constants in interfaces (mspirkov)
 - Enh #349: Add inheritance of trait constants in classes (mspirkov)
 - Enh #261: Add inheritance of interface constants in classes (mspirkov)
+- Chg #352: Remove deprecated code (mspirkov)
 
 
 3.0.8 November 24, 2025